Moved the collectMetadata methods to their own file called TileLayerMetadataWrapper
